Upgrade dagre to 2.0.0
Also ensure explorers are always imported lazily
this reduces the main bundle size

Fixes: #2651
