+++ frontend/biome.json
@@ -60,6 +60,15 @@
                 "useValidTypeof": "error"
             },
             "style": {
+                "noRestrictedImports": {
+                    "level": "error",
+                    "options": {
+                        "paths": {
+                            "@/components/explorer/internal-explorer": "Use Explorer from @/components/explorer/ixplorer instead.",
+                            "@/components/global-explorer/internal-explorer": "Use FullExplorer from @/components/global-explorer/full-explorer instead."
+                        }
+                    }
+                },
                 "noNamespace": "error",
                 "useAsConstAssertion": "error",
                 "useBlockStatements": "off",
