+++ frontend/package.json
@@ -24,7 +24,7 @@
     "dependencies": {
         "@ant-design/charts": "^2.6.7",
         "@ant-design/icons": "^6.1.1",
-        "@dagrejs/dagre": "^1.1.8",
+        "@dagrejs/dagre": "^2.0.4",
         "@fontsource/inter": "^5.2.8",
         "@posthog/react": "^1.8.2",
         "@reduxjs/toolkit": "^2.11.2",
