+++ frontend/src/components/global-explorer/full-explorer.tsx
@@ -1,4 +1,3 @@
-import { ReactFlowProvider } from '@xyflow/react';
 import { lazy, Suspense } from 'react';
 import { LoadingSpinner } from '@/components/loading/loading-spinner/loading-spinner.tsx';
 
@@ -6,10 +5,8 @@ const InternalFullExplorer = lazy(() => import('@/components/global-explorer/int
 
 export function FullExplorer() {
     return (
-        <ReactFlowProvider>
-            <Suspense fallback={<LoadingSpinner />}>
-                <InternalFullExplorer />
-            </Suspense>
-        </ReactFlowProvider>
+        <Suspense fallback={<LoadingSpinner />}>
+            <InternalFullExplorer />
+        </Suspense>
     );
 }
