+++ frontend/src/pages/data-output/components/data-output-tabs/data-output-tabs.tsx
@@ -5,7 +5,7 @@ import { useTranslation } from 'react-i18next';
 import { useLocation, useNavigate } from 'react-router';
 
 import datasetOutlineIcon from '@/assets/icons/dataset-outline-icon.svg?react';
-import { Explorer } from '@/components/explorer/explorer';
+import { Explorer } from '@/components/explorer/explorer.tsx';
 import { HistoryTab } from '@/components/history/history-tab';
 import { LoadingSpinner } from '@/components/loading/loading-spinner/loading-spinner.tsx';
 import { TabKeys } from '@/pages/data-output/components/data-output-tabs/data-output-tabkeys.ts';
